Level 1 scene creation with all camera features implemented+ Player prefab camera removal
Created scene "Level 1" and made a very rough mock-up of a level.

Implemented all camera features, including player following, cutscenes, and camera confiners.

Deleted the Camera object from the Player prefab as it was conflicting with the State Driven Camera used for the cutscene creation. In it's place, there is a new Virtual Cam (called "PlayerCamera").